+++ guniflask/web/__init__.py
@@ -5,6 +5,7 @@ from .blueprint_post_processor import *
 from .context import *
 from .cors import *
 from .global_singleton_annotation import *
+from .param_annotation import *
 from .scheduling_config import *
 
 __all__ = (bind_annotation.__all__ +
@@ -12,4 +13,5 @@ __all__ = (bind_annotation.__all__ +
            context.__all__ +
            cors.__all__ +
            global_singleton_annotation.__all__ +
+           param_annotation.__all__ +
            scheduling_config.__all__)