+++ tests/test_utils/test_factory_utils.py
@@ -18,6 +18,7 @@ class Student(Person):
     secret: str
     mentor: Teacher
     parents: List[Person]
+    scores: dict
 
     def __init__(self):
         self.hobbies: List = None
@@ -26,12 +27,15 @@ class Student(Person):
 
 def test_instantiate_from_json():
     assert instantiate_from_json('1', dtype=int) == 1
+    assert instantiate_from_json([1, 2, 3]) == [1, 2, 3]
+    assert instantiate_from_json({'key': 'value'}) == {'key': 'value'}
 
     student_data = dict(
         name='Xiao Ming',
         age=12,
         graduated=True,
         hobbies=['Programming', 'Piano'],
+        scores={'Math': 100},
         parents=[
             dict(
                 name='Wang Shu',
@@ -56,6 +60,7 @@ def test_instantiate_from_json():
     assert student.age == 12
     assert student.graduated is True
     assert student.hobbies == ['Programming', 'Piano']
+    assert student.scores == {'Math': 100}
     assert student.secret is None
 
     assert isinstance(student.parents, list) and len(student.parents) == 2